Change wall description & add two types of walls
In levels.json there was not very easy wall description (array for each of
wall parameters). Change it to two-dimensional array, now all wall
parameters are in inner array.

Added two types of wall instead of one: horizontal and vertical. The
difference is that they have different friction.

Changed Goals friction to 0.